+++ src/vdj_labeler/CMakeLists.txt
@@ -26,8 +26,8 @@ add_executable(vdj_labeler
                 vdj_alignments/gene_database.cpp
                 vdj_alignments/alignment_structs.cpp
                 vdj_alignments/vj_alignment_info.cpp
-                vdj_alignments/right_v_segment_tail_aligner.cpp
-                vdj_alignments/left_j_tail_aligner.cpp
+                vdj_alignments/aligners/right_v_tail_aligner.cpp
+                vdj_alignments/aligners/left_j_tail_aligner.cpp
                 recombination_calculator/hc_model_based_recombination_calculator.cpp
                 recombination_generation/base_hc_recombination_generator.cpp
                 main.cpp)
