+++ src/vdj_labeler/vdj_alignments/aligners/gene_segment_aligner.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "alignment_structs.hpp"
+#include "../alignment_structs.hpp"
 
 class GeneSegmentAligner {
 public:
