+++ src/vdj_labeler/vdj_alignments/aligners/left_j_tail_aligner.hpp
